+++ RELEASES.md
@@ -95,3 +95,13 @@ Based on:
 - [typescript v0.0.1-alpha.19] .
 ### Releases
 - [NPM v0.0.1-alpha.19] https://www.npmjs.com/package/@gbg/go-core/v/0.0.1-alpha.19 - .
+
+## 2026-03-30 13:50:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.760.0 (2.873.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.1-alpha.20] .
+### Releases
+- [NPM v0.0.1-alpha.20] https://www.npmjs.com/package/@gbg/go-core/v/0.0.1-alpha.20 - .